pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ajib"
dynamic = ["version"]
description = "AJIB — Modular Telegram Bot for VPN customers, resellers, and admins"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "AJIB Contributors" }
]
keywords = ["telegram", "bot", "vpn", "aiogram", "payments", "crypto"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Framework :: AsyncIO",
  "Topic :: Internet",
  "Topic :: System :: Networking",
]

dependencies = [
  "aiogram>=3.3,<4",
  "httpx>=0.27,<1",
  "PyYAML>=6,<7",
  # uvloop only on Linux
  "uvloop>=0.19; platform_system=='Linux'",
]

[project.optional-dependencies]
dev = [
  "black>=23",
  "ruff>=0.5",
  "mypy>=1.6",
  "pytest>=7",
  "pytest-asyncio>=0.23",
  "types-PyYAML>=6",
]

[project.scripts]
# Dev-only runner for the bot (systemd/CLI used in production)
ajib-bot = "ajib.bot.core.bot:main"

[project.urls]
Homepage = "https://github.com/SeyedHashtag/ajib"
Repository = "https://github.com/SeyedHashtag/ajib"
Issues = "https://github.com/SeyedHashtag/ajib/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = { attr = "ajib.__version__" }

[tool.setuptools.packages.find]
where = ["."]
include = ["ajib*"]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
"ajib.bot.locales" = ["*.yml"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I"]
ignore = ["E203"]  # compatible with Black
